% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helper.R
\docType{data}
\name{df}
\alias{df}
\title{Check data frames}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
df
}
\arguments{
\item{d}{a data frame}
}
\description{
This function checks data frames. In seqwrap certain patterns
are assumed, e.g. first column target identifications and following
columns of the same class.
}
\keyword{internal}
