% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap_mtf.R
\name{fit_fun_lme}
\alias{fit_fun_lme}
\title{Special fitting function for lme}
\usage{
fit_fun_lme(fun, arg, vars = NULL)
}
\arguments{
\item{fun}{The fitting function, nlme::lme}

\item{arg}{A list of arguments that can be evaluated by the fitting function}

\item{vars}{A list of variables to be used in the fitting function}
}
\description{
Special fitting function for lme
}
\keyword{internal}
