% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{geo_data}
\alias{geo_data}
\title{Download format gene expression data from Gene Expression Omnibus}
\usage{
geo_data(dir = "./data-raw")
}
\arguments{
\item{dir}{Directory for downloading temporary files. Defaults to "data-raw"}

\item{gse}{GSE accession number}
}
\description{
Currently, a data set from Shetty et al. is made available for the user
after downloading from the Gene Expression Omnibus. The dataset can be used
for replicating analyses in the Vignette and upcoming paper.
}
\references{
Shetty AC, Sivinski J, Cornell J, McCracken C et al. Peripheral
blood transcriptomic profiling of molecular mechanisms commonly
regulated by binge drinking and placebo effects.
\emph{Sci Rep} 2024 May 10;14(1):10733. PMID: 38730024
}
