% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{seqwrap_check}
\alias{seqwrap_check}
\title{Check swcontainer objects for seqwrap}
\usage{
seqwrap_check(x, verbose = TRUE)
}
\arguments{
\item{x}{A swcontainer object}

\item{verbose}{Logical, should the function print diagnostics? Default TRUE}
}
\value{
A list of diagnostics
}
\description{
This function performs verbose checks/diagnostics of a swcontainer object
}
