% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap_mtf.R
\name{seqwrap_mtf}
\alias{seqwrap_mtf}
\title{Transforma and merge data and fit models. The function is used inside
seq_wrapper to combine metadata with target-level data and perform the
model fitting.
The function is used in a call to pbapply::pblapply.}
\usage{
seqwrap_mtf(
  x,
  samp_name,
  metdat,
  arg_list,
  add_vars,
  mt_summary_fun,
  mt_eval_fun,
  ffun,
  return_mod,
  save_mods,
  mod_path
)
}
\arguments{
\item{x}{A data frame of target-specific quantities, in seqwrap a list
from create_list is used iteratively.}

\item{samp_name}{Sample names from the upper level function}

\item{metdat}{Metadata from the upper level function}

\item{arg_list}{Arguments from the upper level function}

\item{add_vars}{Additional variables to keep from the metadata}

\item{mt_summary_fun}{Summary function from the upper level function}

\item{mt_eval_fun}{Evaluation function from the upper level function}

\item{ffun}{the fitting function from the upper level function}

\item{save_mods}{Logical, should the models be saved?}

\item{mod_path}{Path to save the models}
}
\value{
_mod Logical, should the models be returned as part of the results?
}
\description{
Transforma and merge data and fit models. The function is used inside
seq_wrapper to combine metadata with target-level data and perform the
model fitting.
The function is used in a call to pbapply::pblapply.
}
\keyword{internal}
