% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{seqwrapResults}
\alias{seqwrapResults}
\title{seqwrapResults class}
\arguments{
\item{.data}{A list containing initial data (inherited from S7::class_list)}

\item{models}{A list of fitted model objects}

\item{summaries}{A list of model summaries}

\item{evaluations}{A list of model evaluations}

\item{errors}{A data frame of errors and warnings. defaults to an empty dataframe}

\item{n}{Number of samples}

\item{k}{Number of targets}

\item{call_arguments}{Character string of function arguments used}

\item{call_engine}{Character string of modeling engine used}

\item{elapsed_time}{An proc.time() object giving the time needed to complete
iterative model fitting.}
}
\value{
An S7 object of class \code{seqwrap_results} storing fitted models,
summaries, evaluations, and diagnostic information from a \code{seqwrap()} run.
}
\description{
seqwrapResults constructor function.
}
\examples{
# seqwrapResults is the S7 class returned by seqwrap(). End users do
# not normally call the constructor directly -- it is invoked
# internally once iterative model fitting has finished.

library(seqwrap)

dat <- simcounts(n_genes = 5)

container <- seqwrap_compose(
  modelfun = stats::lm,
  arguments = list(formula = y ~ x),
  data = dat$data,
  metadata = dat$metadata,
  samplename = "sample"
)

results <- seqwrap(container, subset = 1:2, cores = 1)

# The object returned by seqwrap() is a seqwrapResults instance
S7::S7_inherits(results, seqwrapResults)

}
