% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{swcontainer}
\alias{swcontainer}
\title{seqwrap container class}
\usage{
swcontainer(
  .data = list(),
  modelfun = NULL,
  arguments = NULL,
  data = NULL,
  rownames = logical(0),
  metadata = structure(list(), names = character(0), row.names = integer(0), class = "data.frame"),
  targetdata = NULL,
  samplename = character(0),
  additional_vars = character(0),
  summary_fun = NULL,
  eval_fun = NULL,
  exported = list(),
  model_print = character(0),
  arguments_print = character(0)
)
}
\arguments{
\item{.data}{A list containing initial data (inherited from S7::class_list)}

\item{modelfun}{A function used to fit models}

\item{arguments}{A list of arguments for the fitting function}

\item{data}{A data frame or list of data frames with target data}

\item{rownames}{Logical, should row names be used as target IDs?}

\item{metadata}{A data frame with sample information}

\item{targetdata}{A data frame with target-wise information}

\item{samplename}{Character for sample name identification}

\item{additional_vars}{Character vector of additional variables}

\item{summary_fun}{A function for summarizing models}

\item{eval_fun}{A function for evaluating models}

\item{exported}{A list of objects to export to workers}

\item{model_print}{Character representation of model function}

\item{arguments_print}{Character representation of arguments}
}
\description{
The seqwrap container is used to store and validate data input
to the to seqwrap.
}
