% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helper.R
\name{data_helper}
\alias{data_helper}
\title{Data helper function.}
\usage{
data_helper(dat, targetdat = NULL, rownames = FALSE)
}
\arguments{
\item{dat}{Sample/target-wise data, either a data frame or a named list.}

\item{targetdat, }{target-wise data, e.g. targe-wise dispersion}
}
\value{
A list of lists, each element of the list has a target specific
data frame over all samples and a target-specific data frame with values
specific to target.
}
\description{
The function takes a named list of data frames
and combines them into target-wise data frames
for use in subsequent modelling.
}
\details{
If the input is a data frame, the function returns a
a list of data frames with a single row (y in subsequent modelling)
}
\keyword{internal}
