% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-summaries.R
\name{generic_evaluation}
\alias{generic_evaluation}
\title{Generic evaluation for model fits}
\usage{
generic_evaluation(x)
}
\arguments{
\item{x}{A model fitted in seqwrap.}
}
\value{
A tidy data frame possible to combine using seqwrap_summarise
}
\description{
When no evaluation function is provided to seqwrap (eval_fun), this function
uses DHARMa to simulate residuals and test for uniformity, dispersion and
outliers. These tests could guide troubleshooting and detection of model
misspecification.
}
