% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-summaries.R
\name{generic_summary}
\alias{generic_summary}
\title{Generic summaries for model parameter estimates.}
\usage{
generic_summary(x)
}
\arguments{
\item{x}{A model fitted in seqwrap.}
}
\value{
A tidy data frame possible to bind using seqwrap_summarise
}
\description{
When no summary function is provided to seqwrap (summary_fun), this function
uses broom(.mixed)::tidy to give a table of model parameter estimates.
}
