% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{print.seqwrapResults}
\alias{print.seqwrapResults}
\title{Print method for objects of class seqwrapResults}
\arguments{
\item{x}{A seqwrapResults object}
}
\value{
Invisibly returns the object
}
\description{
Invoking the print method on seqwrapResults gives a summary
of the fitted objects.
}
\examples{
\dontrun{
results <- seqwrap(...)
print(results)
}
}
