% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{seqwrap_compose}
\alias{seqwrap_compose}
\title{Compose a swcontainer object for use in the seqwrap function.}
\usage{
seqwrap_compose(
  x = NULL,
  modelfun,
  arguments,
  data,
  rownames = FALSE,
  metadata,
  targetdata = NULL,
  samplename = "seq_sample_id",
  additional_vars = NULL,
  summary_fun = NULL,
  eval_fun = NULL,
  exported = list(),
  update = list()
)
}
\arguments{
\item{x}{An optional named list or DGEList object (DESeqDataSet not yet
implemented).}

\item{modelfun}{A model fitting function like stats::lm,
glmmTMB::glmmTMB or lme4::lmer}

\item{arguments}{An alist or list of arguments to be passed to the fitting
function, this should not contain data. Note that the formula must have
y as the dependent variable.}

\item{data}{A data frame or a list of data frames with targets (e.g. genes,
transcripts) as rows and sample names as columns.
If rownames = FALSE (default), each data frame should have target
identifications as the first column in the data frame(s). If rownames = TRUE
row names will be converted to target identifications. If data is provided as
a list, each element of the list should be named. The corresponding names
be available as variables for the fitting function.}

\item{rownames}{should row names in data be used as target identifications?
Defaults to FALSE.}

\item{metadata}{A data frame with sample names (corresponding to column
names in the target matrix)
and design variables.}

\item{samplename}{A character value indicating the variable by which
metadata can merge with the target data. This defaults to "seq_sample_id"
as this is used in the trainomeMetaData package.}

\item{additional_vars}{A vector of additional variables that is contained
in the metadata data set that is needed to fit the model. By default the
metadata is reduced to variables contained in the slots
formula/model/fixed/random in additional arguments.
More variables may be needed for offsets, weights etc.}

\item{summary_fun}{A custom (user-created) function for
evaluating/summarizing models. If NULL, a list of fitted models are returned}

\item{eval_fun}{A custom (user-created) function for model
diagnostics/evaluation. If NULL, no evaluation/diagnostics of models are
returned}

\item{exported}{A list of functions, values etc. to be passed to
summary_fun and eval_fun. This list must contain any functions that
should be used in model summarise or evaluations.}

\item{update}{A list of named parameters to update a swcontainer object.}
}
\value{
A swcontainer object for direct use in seqwrap.
}
\description{
This function makes it possible to run checks on combined data sets
(meta data and target data) and fitting functions to avoid issues
in iterative modelling.
}
