% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{seqwrap_summarise}
\alias{seqwrap_summarise}
\title{Summarise seqwrapResults objects}
\usage{
seqwrap_summarise(
  x,
  summaries = TRUE,
  evaluations = TRUE,
  errors = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{A seqwrapResults object}

\item{summaries}{Logical, should summaries be combined?}

\item{evaluations}{Logical, should evaluations be combined?}

\item{errors}{Logical, should errors be combined?}
}
\description{
Summarise seqwrapResults objects
}
\details{
This functions attempts to summarise results from the summary and evaluation
functions applied in each iteration during modelling. The function expects
that the summary and evaluation functions return data frames.
}
