% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{seqwrapResults}
\alias{seqwrapResults}
\title{seqwrapResults class}
\arguments{
\item{.data}{A list containing initial data (inherited from S7::class_list)}

\item{models}{A list of fitted model objects}

\item{summaries}{A list of model summaries}

\item{evaluations}{A list of model evaluations}

\item{errors}{A data frame of errors and warnings. defaults to an empty dataframe}

\item{n}{Number of samples}

\item{k}{Number of targets}

\item{call_arguments}{Character string of function arguments used}

\item{call_engine}{Character string of modeling engine used}

\item{elapsed_time}{An proc.time() object giving the time needed to complete
iterative model fitting.}
}
\description{
seqwrapResults constructor function.
}
