% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{seqwrapResults}
\alias{seqwrapResults}
\title{seqwrapResults class}
\usage{
seqwrapResults(
  .data = list(),
  models = NULL,
  summaries = NULL,
  evaluations = NULL,
  errors = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  n = integer(0),
  k = integer(0),
  call_arguments = character(0),
  call_engine = character(0)
)
}
\description{
seqwrapResults constructor function.
}
\section{Slots}{

\describe{
\item{\code{models}}{description}

\item{\code{summaries}}{description}

\item{\code{evaluations}}{description}

\item{\code{errors}}{description}

\item{\code{n}}{description}

\item{\code{k}}{description}

\item{\code{call_arguments}}{description}

\item{\code{call_engine}}{description}
}}

