% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{seqwrapResults}
\alias{seqwrapResults}
\title{seqwrapResults class}
\usage{
seqwrapResults(
  .data = list(),
  models = NULL,
  summaries = NULL,
  evaluations = NULL,
  errors = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  n = integer(0),
  k = integer(0),
  call_arguments = character(0),
  call_engine = character(0)
)
}
\arguments{
\item{models}{A list of fitted model objects}

\item{summaries}{A list of model summaries}

\item{evaluations}{A list of model evaluations}

\item{errors}{A data frame of errors and warnings}

\item{n}{Number of samples}

\item{k}{Number of targets}

\item{call_arguments}{Character string of function arguments used}

\item{call_engine}{Character string of modeling engine used}

\item{...}{Additional arguments (for S7 compatibility)}
}
\description{
seqwrapResults constructor function.
}
