% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcounts2.R
\name{simcounts2}
\alias{simcounts2}
\title{Simulate counts from a parallell groups design with three time-points
and two groups.}
\usage{
simcounts2(
  n1 = 5,
  n2 = 5,
  beta0 = c(2.3, 3),
  conditioncon = c(0.2, 0.1),
  timetime2 = c(0, 0),
  timetime3 = c(0.5, 0.25),
  conditioncon_timetime2 = c(0.1, 0.2),
  conditioncon_timetime3 = c(0.5, 0.6)
)
}
\arguments{
\item{n1}{Number clusters in group 1 (control)}

\item{n2}{Number of clusters in group 2 (treatment)}

\item{b0}{Vector sd of the}

\item{b1}{Vector}

\item{b2}{Vector}
}
\description{
Simulate counts from a parallell groups design with three time-points
and two groups.
}
