% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqwrap.R
\name{swcontainer}
\alias{swcontainer}
\title{seqwrap container class}
\usage{
swcontainer(
  .data = list(),
  modelfun = NULL,
  arguments = NULL,
  data = NULL,
  rownames = logical(0),
  metadata = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  targetdata = NULL,
  samplename = character(0),
  additional_vars = character(0),
  summary_fun = function() NULL,
  eval_fun = function() NULL,
  exported = list(),
  model_print = character(0)
)
}
\description{
The seqwrap container is used to store and validate data input
to the to seqwrap.
}
\section{Slots}{

\describe{
\item{\code{modelfun}}{A function used to fit models using data and meta data
as input.}

\item{\code{arguments}}{A list of arguments provided to the fitting function.}

\item{\code{data}}{A data frame or a list of data frames}

\item{\code{rownames}}{A logical indicating if row names should be used as target id}

\item{\code{metadata}}{A data frame with sample information}

\item{\code{targetdata}}{A data frame with sample information}

\item{\code{samplename}}{A character for sample name identification}

\item{\code{additional_vars}}{A character vector of additional variables exported to}

\item{\code{summary_fun}}{A function}

\item{\code{eval_fun}}{A function}

\item{\code{exported}}{A list}
}}

